+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+vNext
+  * Add `Model.bulkCreate()` support
+
 v0.6.1 - d65cbc9 - Dec 7th 2016
   * Fix `Instance` initialization modifying the original passed in object